task-2300- search by filter terms
